+++ CHANGELOG.md
@@ -5,10 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-## [6.2.1] - 2023-03-23
 ### Added
 - Add `hideToolbar` to resume & JD search config
+- Add ExtractorConfig object to Collection
 
 ## [6.2.1] - 2023-03-23
 ### Added